#78 Refactor the code (#132)
