+++ FatNetLib.Tests/Modules/Json/PackageSchemaConverterTests.cs
@@ -3,7 +3,7 @@ using Kolyhalov.FatNetLib.Initializers;
 using Newtonsoft.Json;
 using NUnit.Framework;
 
-namespace Kolyhalov.FatNetLib;
+namespace Kolyhalov.FatNetLib.Modules.Json;
 
 public class PackageSchemaConverterTests
 {
