+++ FatNetLib/FatNetLib.csproj
@@ -20,7 +20,10 @@
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="StyleCopPlus" Version="3.0.0"/>
+        <PackageReference Include="StyleCopPlus" Version="3.0.0">
+            <PrivateAssets>all</PrivateAssets>
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
         <PackageReference Include="BouncyCastle.NetCore" Version="1.9.0"/>
     </ItemGroup>
     <ItemGroup>
