+++ FatNetLib/Initializers/Controllers/Server/ExchangeEndpointsController.cs
@@ -17,7 +17,6 @@ public class ExchangeEndpointsController : IController
 
     [Route("exchange")]
     [Schema(key: nameof(Package.Body), type: typeof(EndpointsBody))]
-    [return: Schema(key: nameof(Package.Body), type: typeof(EndpointsBody))]
     public Package ExchangeEndpoints(Package handshakePackage)
     {
         int clientPeerId = handshakePackage.FromPeerId!.Value;
