+++ project.dic
@@ -5,6 +5,7 @@ marocchino
 ntrubkin
 nupkg
 kzrnm
+decryptor
 эндпоинт
 эндпоинта
 коннекшн
